orchestrator: increase ConditionUnknown timeout
Orchestrator does periodic health checking (discoveries) for every known
MySQL instance. This is driven in normal situation by [healthTick][ht]
firing by default every second. The health tick adds instances which have
`IsUpToDate == false` to the discovery queue. The queue is then consumed by
the discovery routine which updates the instances, but only those with
`IsUpToDate == false` ([second check][disc]).

So the instance healthcheck is driven by IsUpToDate property, and there
is a window where the instance is not actually up to date and is beeing
checked or will be soon scheduled to be.

At the same time the operator uses this property to identify instanaces
with outdated checks and configures them with unknown status. This race
leads to periodic switches of MySQL pods from services.

Property `IsRecentlyChecked` is the [five times][rece] the basic
`InstancePollSeconds` and should be more representative of Unknown state
for MySQL instance.

[ht]: https://github.com/openark/orchestrator/blob/master/go/logic/orchestrator.go#L342
[disc]: https://github.com/openark/orchestrator/blob/master/go/logic/orchestrator.go#L231
[oper]: https://github.com/presslabs/mysql-operator/blob/master/pkg/controller/orchestrator/orchestrator_reconcile.go#L211
[rece]: https://github.com/openark/orchestrator/blob/761be91299ce81abf6c9a3889a660549b8daef10/go/inst/instance_dao.go#L1153