sidecar: add missing exec to entrypoint
If sidecar container receives SIGTERM the bash will not properly
terminate the sidecar agent. Since the shell is not performing any
function after the sidecar is launched, just exec to sidecar.

This way the SIGTERM is received directly by sidecar terminating
imeddiately and the pod doesn't wait for grace termination period.