pt-heartbeat: force UTC timezone
it's possible to create a MySQL Cluster instance with non default
timezone `default-time-zone: Somewhere\Something`. In which case
the Orchestrator will read up replication lag in adjusted time and assumes
the node have a replication lag (or comes from future).

Forcing UTC on both write and read endpoints to work correctly even
in this condition.