fix orchestrator startup crash in DualStack cluster
In a cluster where `.Env.HOSTNAME` resolves to an IPv6 address, the
orchestrator fails to start due to a syntax error in the raft bind
address. This was reported here:
https://github.com/bitpoke/mysql-operator/issues/929

This commit fixes this, at least for DualStack clusters, by using
`.Env.POD_IP` instead, which is defined as

```yaml
- name: POD_IP
  valueFrom:
    fieldRef:
      apiVersion: v1
      fieldPath: status.podIP
```

and thus typically points to an IPv4 address in DualStack clusters.
For a general fix, the orchestrator code itself probably needs to
be modified to correctly handle the bind hostname.